+++ packages/plugin-image/web/src/image-viewer.jsx
@@ -9,6 +9,7 @@ import {
   isSSR,
   getImageSrc,
   WIX_MEDIA_DEFAULT,
+  anchorScroll,
 } from 'wix-rich-content-common';
 // eslint-disable-next-line max-len
 import pluginImageSchema from 'wix-rich-content-common/dist/statics/schemas/plugin-image.schema.json';
@@ -221,7 +222,7 @@ class ImageViewer extends React.Component {
       },
     } = this.props;
     const element = document.getElementById(`viewer-${anchor}`);
-    element.scrollIntoView({ behavior: 'smooth' });
+    anchorScroll(element);
   };
 
   hasLink = () => this.props.componentData?.config?.link?.url;