+++ packages/plugin-link/web/src/LinkViewer.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { normalizeUrl, mergeStyles, validate } from 'wix-rich-content-common';
+import { normalizeUrl, mergeStyles, validate, anchorScroll } from 'wix-rich-content-common';
 import pluginLinkSchema from 'wix-rich-content-common/dist/statics/schemas/plugin-link.schema.json';
 import { isEqual } from 'lodash';
 import styles from '../statics/link-viewer.scss';
@@ -36,7 +36,7 @@ class LinkViewer extends Component {
     this.props?.settings?.onClick?.(event, anchor || this.getHref());
     if (anchor && !isInEditor) {
       const element = document.getElementById(`viewer-${anchor}`);
-      element.scrollIntoView({ behavior: 'smooth' });
+      anchorScroll(element);
     }
   };