+++ packages/plugin-link/web/src/toolbar/UrlLinkButton.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { getLinkDataInSelection } from 'wix-rich-content-editor-common';
 import styles from '../../statics/link-viewer.scss';
-import { normalizeUrl, mergeStyles } from 'wix-rich-content-common';
+import { normalizeUrl, mergeStyles, anchorScroll } from 'wix-rich-content-common';
 
 export default class UrlLinkButton extends Component {
   constructor(props) {
@@ -19,7 +19,7 @@ export default class UrlLinkButton extends Component {
     const nodeListOfAllblocks = document.querySelectorAll(`[data-editor]`);
     const arrayOfAllblocks = Array.apply(null, nodeListOfAllblocks);
     const element = arrayOfAllblocks.find(block => block.dataset.offsetKey === `${anchor}-0-0`);
-    element.scrollIntoView({ behavior: 'smooth' });
+    anchorScroll(element);
   };
 
   preventDefault = event => event.preventDefault();